Move yaml/yaml.h to yaml.h and merge version.c to api.c.
